refractor: cleaner type
