--- app/src/main/java/org/hospitalmanager/model/RefreshTokenRequest.java
@@ -1,13 +1,0 @@
-package org.hospitalmanager.model;
-
-public class RefreshTokenRequest {
-    private String refreshToken;
-
-    public RefreshTokenRequest(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-}
