--- app/src/main/java/org/hospitalmanager/model/RefreshTokenResponse.java
@@ -1,19 +1,0 @@
-package org.hospitalmanager.model;
-
-public class RefreshTokenResponse {
-    private String accessToken;
-    private String refreshToken;
-
-    public RefreshTokenResponse(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-}
