+++ app/src/main/java/org/hospitalmanager/model/ResetPassword.java
@@ -1,0 +1,23 @@
+package org.hospitalmanager.model;
+
+public class ResetPassword {
+    public static class Request {
+        private String email;
+
+        public String getEmail() {
+            return email;
+        }
+    }
+
+    public static class Response {
+        private String message;
+
+        public Response(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+}
