--- app/src/main/java/org/hospitalmanager/model/ResetPasswordRequest.java
@@ -1,9 +1,0 @@
-package org.hospitalmanager.model;
-
-public class ResetPasswordRequest {
-    private String email;
-
-    public String getEmail() {
-        return email;
-    }
-}
