--- app/src/main/java/org/hospitalmanager/model/ResetPasswordResponse.java
@@ -1,13 +1,0 @@
-package org.hospitalmanager.model;
-
-public class ResetPasswordResponse {
-    private String message;
-
-    public ResetPasswordResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
